refactor: move open-exchange client logic into internal/daemon
